fixing sign conversion warning
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@38844 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/mac/carbon/window.cpp
@@ -507,7 +507,7 @@ pascal OSStatus wxMacUnicodeTextEventHandler( EventHandlerCallRef handler , Even
         numChars = dataSize / sizeof( UniChar) + 1;
         charBuf = buf ;
 
-        if ( numChars * 2 > sizeof(buf) )
+        if ( (size_t) numChars * 2 > sizeof(buf) )
             charBuf = new UniChar[ numChars ] ;
         else
             charBuf = buf ;
